--- -ms-empresa/src/main/java/com/api/empresa/EmpresaApplication.java
@@ -1,13 +1,0 @@
-package com.api.empresa;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EmpresaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EmpresaApplication.class, args);
-	}
-
-}
