--- -ms-empresa/src/main/java/com/api/empresa/repositorio/SetorRepositorio.java
@@ -1,8 +1,0 @@
-package com.api.empresa.repositorio;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.api.empresa.entidade.Setor;
-
-public interface SetorRepositorio extends JpaRepository<Setor, Long> {
-}
