--- -ms-empresa/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-spring.application.name=empresa
-spring.datasource.url=jdbc:mysql://localhost:3306/API3?useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-# JPA Properties
-spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
