--- -ms-empresa/src/test/java/com/api/empresa/EmpresaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.api.empresa;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EmpresaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
