+++ -ms-usuario/src/main/java/com/fatec/ms_usuario/repositorio/SetorRepositorio.java
@@ -1,0 +1,8 @@
+package com.fatec.ms_usuario.repositorio;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.fatec.ms_usuario.entidade.Setor;
+
+public interface SetorRepositorio extends JpaRepository<Setor, Long> {
+}
