--- ms-usuario/src/main/resources/application.properties
@@ -1,15 +1,0 @@
-spring.application.name=ms-usuario
-
-spring.datasource.url=jdbc:mysql://localhost:3308/des-web
-spring.datasource.username=root
-spring.datasource.password=salerno
-spring.jpa.show-sql=true
-
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.hibernate.ddl-auto=create
-
-# Dialeto do Hibernate
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
-
-# Driver JDBC
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
